requirements.txt
# Core dependencies
openpyxl==3.1.2
python-dotenv==1.0.0
schedule==1.2.0
keyboard==0.13.5
boxsdk==3.9.2

# Optional but recommended for development
pytest==7.4.0
black==23.7.0
flake8==6.0.0